% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs.R
\name{encode_lfs_factors}
\alias{encode_lfs_factors}
\title{Produce a data frame (tibble) with labelled vectors encoded as factors}
\usage{
encode_lfs_factors(records, codebook)
}
\arguments{
\item{records}{A data frame (tibble) of records}

\item{codebook}{A data frame (tibble) of codebook}
}
\value{
A data frame (tibble) with labelled vectors encoded as factors
}
\description{
Produce a data frame (tibble) with labelled vectors encoded as factors
}
