% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs.R
\name{fetch_lfs_pumf}
\alias{fetch_lfs_pumf}
\alias{.fetch_single_year}
\title{Download LFS PUMF data from StatCan website}
\usage{
fetch_lfs_pumf(
  years,
  cache_dir = here::here(),
  refresh_cache = FALSE,
  keep_zip = FALSE,
  extract = TRUE
)

.fetch_single_year(
  year,
  cache_dir,
  refresh_cache = FALSE,
  extract = TRUE,
  keep_zip = FALSE
)
}
\arguments{
\item{years}{A vector of years to download}

\item{cache_dir}{The directory to cache the downloaded files}

\item{refresh_cache}{Whether to force download even if the files are already cached}
}
\value{
Invisible
}
\description{
Download LFS PUMF data from StatCan website
}
