% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs.R
\name{generate_lfs_bootstrap_weights}
\alias{generate_lfs_bootstrap_weights}
\alias{.sample_poisson_factors}
\alias{.generate_replicates}
\alias{.calibrate_weights}
\alias{.calculate_age_tabs}
\title{Generate bootstrap weights for LFS PUMF data}
\usage{
generate_lfs_bootstrap_weights(d, n_reps)

.sample_poisson_factors(k)

.generate_replicates(final_weight, n_reps)

.calibrate_weights(uncalibrated_weights, final_weight, domains)

.calculate_age_tabs(age_6, age_12)
}
\arguments{
\item{d}{:data.frame (labelled) LFS PUMF data.}

\item{n_reps}{:integer number of bootstrap replicates to generate.}

\item{age_6}{:factor AGE_6 variable.}

\item{age_12}{:factor AGE_12 variable.}
}
\value{
a (n_reps by nrow(d)) array of bootstrap weights.

a factor AGE_TABS variable.
}
\description{
Generate bootstrap weights for LFS PUMF data

Sample poisson factors for simulated bootstrap weights

initialize uncalibrated bootstrap replicate weights

Calibrate uncalibrated bootstrap replicate weights

Calculate AGE_TABS for LFS PUMF data
}
