% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs.R
\name{read_lfs_pumf}
\alias{read_lfs_pumf}
\alias{read_lfs_records}
\alias{.correct_decimal_placement}
\alias{read_lfs_codebook}
\alias{encode_lfs_factors}
\title{Read LFS PUMF}
\usage{
read_lfs_pumf(path)

read_lfs_records(path)

.correct_decimal_placement(x)

read_lfs_codebook(path)

encode_lfs_factors(records, codebook)
}
\arguments{
\item{path}{The directory or zip archive containing the LFS PUMF (e.g. data/raw/2020)}

\item{records}{A data frame (tibble) of records}

\item{codebook}{A data frame (tibble) of codebook}
}
\value{
A tbl_df of LFS PUMF records in \code{dir} labelled with the codebook in \code{dir}

A data frame (tibble) with labelled vectors encoded as factors
}
\description{
Read LFS PUMF

Produce a data frame (tibble) with labelled vectors encoded as factors
}
