% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs.R
\name{roxygenize_lfs_codebook}
\alias{roxygenize_lfs_codebook}
\alias{.cb_top}
\alias{.cb_gen}
\alias{.cb_format}
\alias{.cb_column}
\alias{.cb_column__number}
\alias{.cb_column__factor}
\alias{.cb_truncate_factor}
\alias{.cb_source}
\alias{.cb_describe}
\title{Produce data documentation in standard Roxygen markdown}
\usage{
roxygenize_lfs_codebook(records, codebook, year)

.cb_top(year)

.cb_gen(...)

.cb_format(records, codebook)

.cb_column(col, records)

.cb_column__number(col, records)

.cb_column__factor(col)

.cb_truncate_factor(factor_labels)

.cb_source(year)

.cb_describe(year)
}
\arguments{
\item{records}{A data frame (tibble) of records}

\item{codebook}{A data frame (tibble) of codebook}

\item{year}{The year of the data}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> key-value pairs}

\item{col}{a column from the codebook, as a named list with:
\itemize{
\item \code{field_id}: the field ID
\item \code{var_name}: the variable name
\item \code{var_label}: the variable label
\item \code{is_factor}: whether the variable is a factor
\item \code{factor_labels}: if \code{is_factor}, a data frame (tibble) of factor labels
}}

\item{description}{A description of the data}
}
\value{
A character vector of Roxygen markdown
}
\description{
Produce data documentation in standard Roxygen markdown

Produce the top of data documentation

Generate roxygen documentation from key-value pairs

Generate roxygen documentation for the format of a data set

Generate roxygen documentation for a column of the data set

Generate roxygen documentation for a numeric column of the data set

Generate roxygen documentation for a factor column of the data set

Truncate n-level factors to 7 levels
(levels\link{1, 2, 3}, ..., levels\link{n-2, n-1, n})

Generate roxygen documentation for the source of a data set

Generate roxygen documentation for the description of a data set
}
