% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs.R
\name{read_lfs_pumf}
\alias{read_lfs_pumf}
\alias{.read_lfs_records}
\alias{.correct_decimal_placement}
\alias{.read_lfs_codebook}
\title{Read LFS PUMF}
\usage{
read_lfs_pumf(dir)

.read_lfs_records(dir)

.correct_decimal_placement(x)

.read_lfs_codebook(dir)
}
\arguments{
\item{dir}{The directory containing the LFS PUMF (e.g. data/raw/2020)}
}
\value{
A list with two elements: records and codebook
}
\description{
Read LFS PUMF
}
